Tie OriginAccessEntry into WebPublicSuffixList.
Part four of a two-sided patch that prevents setting document.domain
to a top-level domain like "com" or "co.uk".

Also improves the console messages when this test fails. We need to return
more information from OriginAccessEntry to make these better messages.

Adds a unit test for this functionality. Doing this as part of a layout
test is precluded by the lack of actual x.y.com domain names for the
test webserver.

BUG=307407
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/76363007

git-svn-id: svn://svn.chromium.org/blink/trunk@162427 bbb929c8-8fbe-4397-9dbb-9b2b20218538