Revert: Fix animation of points attribute for polygons
This patch reverts https://src.chromium.org/viewvc/blink?view=rev&revision=162399
which was already fixed in Chromium. When I merged this, I didn't notice the
tests passed before the patch was applied.

These tests also have turned flaky on the bots. The current theory is that this
is unrelated to this change.

TBR=schenney@google.com
BUG=321710

Review URL: https://codereview.chromium.org/81413003

git-svn-id: svn://svn.chromium.org/blink/trunk@162492 bbb929c8-8fbe-4397-9dbb-9b2b20218538