CSSGrammary.y: Fix error recovery in grouping rule bodies
Problem with the current grammar is as follows. If a sequence of tokens looks like a rule start, it is reduced by the rule recovery code. It eats any token until it finds ";", "{" or EOF.

To fix that, any sequence of tokens must be reduceable to:
1) invalid_rule, if next token is ";" or "{".
2) invalid_rule_header, if next token is "}".

This CL leaves @import unfixed. It's the only rule that may be valid without ";" or "{". Current grammar produces valid @import rule for the stylesheet "@media all{ @import url("")" (because of the autoclosing curly braces), but adding "}" makes it invalid (because it no longer ends with EOF). So it's a more complicated case.

R=tabatkins@google.com, apavlov@chromium.org

BUG=321540

Review URL: https://codereview.chromium.org/80613005

git-svn-id: svn://svn.chromium.org/blink/trunk@162537 bbb929c8-8fbe-4397-9dbb-9b2b20218538