js-test.js: Make shouldBeFoo(expr, quiet) actually use quiet flag
A handful of tests pass in a |quiet| flag that's not implemented.
Fix it by actually implementing it and updating tests/results.

Also, make use of _-prefixed arguments and variables consistent
in functions that call eval() (to protect them) and don't prefix
them in functions that do not call eval().

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/84123002

git-svn-id: svn://svn.chromium.org/blink/trunk@162583 bbb929c8-8fbe-4397-9dbb-9b2b20218538