Stub out ControllerGlobalScope
This is the first patch for implementing NavigationController. This patch
stubs out the JS scope for the new class of worker. It does not introduce
a new Worker / Thread implementation, merely allows the existing SharedWorkerThread
to host a NavigationController thread.

The threads are created by the backend (browser) and not exposed directly to
script running in a web page.
BUG=

Review URL: https://codereview.chromium.org/21864005

git-svn-id: svn://svn.chromium.org/blink/trunk@162637 bbb929c8-8fbe-4397-9dbb-9b2b20218538