Re-use nameForAutofill between WebFormControlElement and WebPasswordFormData
Autofill and password management code needs to learn a string identifying a form element. This string is preferably the value of the "id" attribute, or, if that's missing, the "name" attribute.

Previously, WebFormControlElement offered nameForAutofill to get this identification, while WebPasswordFormData had a separate function to achieve the same.

This CL moves the nameForAutoFill code up to HTMLFormControlElement, and uses it from there on both places mentioned above.

BUG=322242, 282487

Review URL: https://codereview.chromium.org/85873002

git-svn-id: svn://svn.chromium.org/blink/trunk@162696 bbb929c8-8fbe-4397-9dbb-9b2b20218538