Add test for innerHTML with a wide DOM
We already have a test for testing a "large" DOM
but it only has a single root node, which avoided
some hot paths.

This test is based off of http://jsperf.com/parse-html-type
which was written by a member of the Google+ team who noticed
a regression in our innerHTML performance as of late.

This is a tiny microbenchmark, which happens to catch some
of this regression, but probably isn't one we should
spend a lot of time optimizing (we're already very very fast).

On my machine this clocks at about 6,000 operations per second.

BUG=318251

Review URL: https://codereview.chromium.org/70953002

git-svn-id: svn://svn.chromium.org/blink/trunk@162960 bbb929c8-8fbe-4397-9dbb-9b2b20218538