Remove TreatNullAs=NullString for HTMLInputElement
http://whatwg.org/html#htmlinputelement
http://whatwg.org/html#HTMLInputElement-partial

The value attribute has [TreatNullAs=EmptyString] in the spec, so
leave it with [TreatNullAs=NullString] even though that doesn't
exactly match the spec, to be fixed in a later commit.

The tests for the updated properties pass in Firefox Nightly, IE11
Release Preview and Opera 12.16, with these exceptions:

Firefox reflects autocomplete as the empty string, and doesn't support
dirName.

IE throws an exception for align/type, doesn't support dirName, and
doesn't reflect formAction as a URL.

Opera doesn't reflect formAction as a URL, and reflects min/max as the
empty string.

BUG=310298

Review URL: https://codereview.chromium.org/98383004

git-svn-id: svn://svn.chromium.org/blink/trunk@163033 bbb929c8-8fbe-4397-9dbb-9b2b20218538