Remove TreatNullAs=NullString for HTMLScriptElement
http://whatwg.org/html#htmlscriptelement
http://whatwg.org/html#HTMLScriptElement-partial

The nonce attribute is experimental and not in the HTML spec, but
remove [TreatNullAs=NullString] anyway since it's hardly intended.

The tests for the updated properties pass in Firefox Nightly, IE11
Release Preview and Opera 12.16, with these exceptions:

Firefox does not support nonce.

IE does not support crossOrigin or nonce.

Opera throws an exception when setting text, and does not support
htmlFor, event, crossOrigin or nonce.

BUG=310298

Review URL: https://codereview.chromium.org/98543006

git-svn-id: svn://svn.chromium.org/blink/trunk@163036 bbb929c8-8fbe-4397-9dbb-9b2b20218538