Initial touch-action main thread implementation
This adds initial support for touch-action: none/auto on the blink main thread.
This works by computing the effective touch action on any touchstart event and
reporting the result to the embedder.  Chrome will then do appropriate filtering
of gesture events in the browser process.  This is necessary because the browser may do other optimizations based on the result (such as suppressing sending of
touchmove events during a scroll).  We report this via WebWidgetClient in advance of the
touch event ACK because in the future it will need to be able to come from the impl thread without waiting on main (plus there can be some performance benefit to doing it in advance
of running any JS handlers).

For more details see the design document:
https://docs.google.com/a/chromium.org/document/d/1CV2AXyrdPdGSRypAQcfGrgQVuWYi50EzTmVsMLWgRPM/edit#

Note that this does not yet integrate at all with compositor fast paths (eg.
a touch handler is necessary for touch-action to take effect in order to
bypass compositor touch hit testing).

Intent-to-implement thread: https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/mDSwpxfdrdA

Chromium-side patch to consume this: https://codereview.chromium.org/67383002/

BUG=316735

Review URL: https://codereview.chromium.org/16507017

git-svn-id: svn://svn.chromium.org/blink/trunk@163373 bbb929c8-8fbe-4397-9dbb-9b2b20218538