Web Animations CSS: Fall back to legacy behavior of avoiding cross-fades of generated images.
When interpolating between mismatched gradients (and other generated image types), the correct behavior is to cross-fade. However, cross-fading of generated images is not currently implemented in Blink.

This patch reverts to the legacy CSS animation behavior of switching instantaneously between generated images on animation. We should revert this patch when cross-fading of generated images is supported.

BUG=327124

Review URL: https://codereview.chromium.org/111343002

git-svn-id: svn://svn.chromium.org/blink/trunk@163505 bbb929c8-8fbe-4397-9dbb-9b2b20218538