Merge CSSAnimationUpdateScope into styleForRenderer()
CSSAnimationUpdateScope was hiding some logic and an unneccesary null check
and should only be used in recalcStyle. Remove this class and move the
logic into Element::styleForRenderer() which is what the RAII thing was
always bracketed around.

This also lets NodeRenderingContext become unaware of animations, it doesn't
need to be concerned with that.

Review URL: https://codereview.chromium.org/111003005

git-svn-id: svn://svn.chromium.org/blink/trunk@163568 bbb929c8-8fbe-4397-9dbb-9b2b20218538