Add a missing return statement.
It looks like this was omitted accidentally. This isn't a behavior change
as the next case contains the same if check that the current case contained
(which will not be taken, because it wasn't taken in the current case), and
the next intended state is identical. So no behavior change other than saving
one conditional and making the code less confusing.

BUG=177475

Review URL: https://codereview.chromium.org/103933006

git-svn-id: svn://svn.chromium.org/blink/trunk@163664 bbb929c8-8fbe-4397-9dbb-9b2b20218538